Rename `utils.functions` to `utils.commons` (#351)
