+++ camel/utils/commons.py
